Refactored Home components and SetDetails.tsx
