+++ client/src/components/SetDetails.tsx
@@ -2,7 +2,7 @@ import { Link, useParams } from "react-router-dom";
 import React, { useEffect, useState } from "react";
 import { FlippableCardContainer } from "./flippableCardContainer";
 
-import "./SetDetails.css";
+import "./styles/SetDetails.css";
 
 interface Set {
   _id: string;
